Move sorter test functions into a header
